добавлена проверка наличия файла

+++ src/core/loader.py
@@ -22,8 +22,11 @@ class SchemaLoader:
             return json.loads(f.read(-1))
     
     @classmethod
-    def dump(cls, scheme:str, data:dict) -> None:
-        file = f'{os.getcwd()}/schemes/{scheme}.json'
-        with open(file, 'w') as f:
-            f.write(json.dumps(data))
+    def dump(cls, source:str, scheme:str, data:dict) -> None:
+        file = f'{os.environ["SCHEMES_PATH"]}/{source}/{scheme}.json'
+        if os.path.isfile(file):
+            with open(file, 'w') as f:
+                f.write(json.dumps(data, ensure_ascii=False, indent=4, sort_keys=True))
+        else:
+            raise FileExistsError
